clippy: fix connection to detailed notify signal
g_quark_try_string() will only return a non-0 GQuark when one already
exists for the given string. That will not be the case when
clippy_connect() is called and a GQuark is attempted to be created for
the detail string.

Simply use g_quark_from_string() instead to fix the issue. Since it
returns 0 when NULL is passed-in, the clippy_return_if_fail() will
still return if a detail is not provided.