utils: fix object_get_name() for JS proxies
ClippyJsProxy is neither a GtkWidget nor a GtkBuildable; prior to this
commit, those are the only two cases handled by object_get_name().
The effect is that when a JS proxy is used, signals like ObjectNotify
which rely on the return value of object_get_name() for their details,
will not be emitted with the expected values.

This commit fixes the issue by storing the full name a given proxy is
created for, at time of construction, and by adding a new case to
object_get_name() to handle it.