clippy: wait for the webview to be loaded before executing JS
At the moment, app_get_object() assumes that whichever object is
required in the JS context maps to JS code that has already been
loaded.
However, now that we changed the way Clippy is loaded, that may not
yet have happened before e.g. Connect() is called from a client.

This commit ensures the webview is fully loaded before we create the
ClippyJsProxy.