flatpak-overrides: Allow all apps access to sound server
We need this in order to allow hacked apps to play sounds. Even though
this is not a logical place to put this override, we need to keep the
override in a flatpak so that it can be updated through the software
center, and it needs to be in the same place as the Clippy overrides
because there is only one global overrides file. This seems as good a
place as any.

https://phabricator.endlessm.com/T24260